requirements.txt
APScheduler==3.10.4
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
motor>=3.3.1
beanie>=1.22.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
openpyxl>=3.1.5
reportlab>=4.2.2
pymongo>=4.5.0
cors>=1.0.1
python-dotenv>=1.0.0

